+++ server/controllers/apiQueries.js
@@ -9,19 +9,19 @@ const Medication = require('../models/medication')
 module.exports = {
   // ========== Patient Queries ========
   findAllPatients: () => {
-    return Patient.find({})
+    return Patient.find({}, { password: 0 })
     .populate('appointments')
     .populate('providers')
     .populate('medications')
   },
   findPatientByEmail: (email) => {
-    return Patient.findOne({ email })
+    return Patient.findOne({ email }, { password: 0 })
     .populate('appointments')
     .populate('providers')
     .populate('medications')
   },
   findPatientById: (_id) => {
-    return Patient.find({ _id })
+    return Patient.find({ _id }, { password: 0 })
     .populate('appointments')
     .populate('providers')
     .populate('medications')